src/providers/llms/gemini/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nucleusiq-gemini"
version = "0.1.0"
description = "Google Gemini provider for the NucleusIQ AI agent framework."
requires-python = ">=3.10,<4.0"
license = { text = "MIT" }
authors = [{ name = "Nucleusbox", email = "info@nucleusbox.com" }]
keywords = ["AI", "agents", "gemini", "google", "nucleusiq"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "nucleusiq>=0.6.0",
    "google-genai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/nucleusbox/NucleusIQ"
Repository = "https://github.com/nucleusbox/NucleusIQ"
Issues = "https://github.com/nucleusbox/NucleusIQ/issues"

[dependency-groups]
test = [
    "pytest>=8.0,<10.0",
    "pytest-asyncio>=0.23,<2.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]
dev = [{ include-group = "test" }]

[tool.uv.sources]
nucleusiq = { path = "../../../nucleusiq", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["nucleusiq_gemini"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that make real API calls (require GEMINI_API_KEY)",
]
